Switch from "size_t" to "u32" to improve compatibility across C/C++

+++ include/primes.h
@@ -41,7 +41,7 @@ bool is_prime(u64 num);
 u32* get_prime_fact(u32 num, u16 size);
 u64* get_prime_fact_64(u64 num, u32 size);
 bool* generate_sieve(u32 limit);
-u32* generate_primes(bool* sieve, u32 limit, size_t* count);
+u32* generate_primes(bool* sieve, u32 limit, u32* count);
 
 #ifdef __cplusplus
 }
